Rydder i struktur og filer
