+++ libs/etterlatte-regler/src/test/kotlin/beregning/BeregnBarnepensjon.kt
@@ -1,8 +1,8 @@
 package beregning
 
-import FaktumNode
-import RegelGrunnlag
-import ToDoRegelReferanse
+import regler.FaktumNode
+import regler.RegelGrunnlag
+import regler.ToDoRegelReferanse
 import beregning.barnepensjon1967.BP_1967_DATO
 import beregning.barnepensjon1967.beregnBarnepensjon1967Regel
 import beregning.barnepensjon2024.beregnBarnepensjon2024Regel
