+++ libs/etterlatte-regler/src/test/kotlin/beregning/BeregnBarnepensjonTest.kt
@@ -1,19 +1,19 @@
 package beregning.barnepensjon1967
 
-import FaktumNode
 import beregning.AvdoedForelder
 import beregning.BarnepensjonGrunnlag
 import beregning.beregnBarnepensjonRegel
-import finnAlleKnekkpunkter
 import io.kotest.matchers.shouldBe
 import no.nav.etterlatte.libs.common.grunnlag.Grunnlagsopplysning
 import no.nav.etterlatte.libs.common.toJson
 import org.junit.jupiter.api.Test
+import regler.FaktumNode
+import regler.finnAlleKnekkpunkter
 import java.math.BigDecimal
 import java.time.Instant
 import java.time.YearMonth
 
-class Barnepensjon1967Test {
+class BeregnBarnepensjonTest {
     private val saksbehandler = Grunnlagsopplysning.Saksbehandler("Z12345", Instant.now())
     private val grunnlag = BarnepensjonGrunnlag(
         grunnbeloep = FaktumNode(kilde = saksbehandler, beskrivelse = "Grunnbeløp", verdi = BigDecimal(100_550)),
