+++ libs/etterlatte-regler/src/test/kotlin/beregning/barnepensjon1967/BeregnBarnepensjon1967.kt
@@ -1,10 +1,10 @@
 package beregning.barnepensjon1967
 
-import ToDoRegelReferanse
 import beregning.BarnepensjonGrunnlag
 import beregning.barnepensjon1967.barnekull.barnekullRegel
 import beregning.barnepensjon1967.trygdetidsfaktor.trygdetidsFaktor
 import regler.RegelMeta
+import regler.ToDoRegelReferanse
 import regler.definerKonstant
 import regler.kombinerer
 import regler.med
