+++ libs/etterlatte-regler/src/test/kotlin/beregning/barnepensjon2024/BeregnBarnepensjon2024.kt
@@ -1,6 +1,6 @@
 package beregning.barnepensjon2024
 
-import ToDoRegelReferanse
+import regler.ToDoRegelReferanse
 import beregning.BarnepensjonGrunnlag
 import beregning.barnepensjon1967.BP_1967_DATO
 import beregning.barnepensjon1967.kroneavrundingKonstant
@@ -15,9 +15,9 @@ import java.time.LocalDate
 
 val BP_2024_DATO: LocalDate = LocalDate.of(2024, 1, 1)
 
-val sats2024 = definerKonstant<BarnepensjonGrunnlag, BigDecimal>(
+val beloepEnForelderDoed = definerKonstant<BarnepensjonGrunnlag, BigDecimal>(
     gjelderFra = BP_2024_DATO,
-    beskrivelse = "Prosentsats benyttet for første barn",
+    beskrivelse = "Kronebelop benyttet for barn med en forelder død",
     regelReferanse = ToDoRegelReferanse(),
     verdi = BigDecimal(100_000)
 )
@@ -26,7 +26,7 @@ val reduksjonMotFolketrygdRegel = RegelMeta(
     gjelderFra = BP_1967_DATO,
     beskrivelse = "Reduserer ytelsen mot opptjening i folketrygden",
     regelReferanse = ToDoRegelReferanse()
-) kombinerer sats2024 og trygdetidsFaktor med { sats, trygdetidsfaktor ->
+) kombinerer beloepEnForelderDoed og trygdetidsFaktor med { sats, trygdetidsfaktor ->
     (sats * trygdetidsfaktor)
 }
